:sparkles: add RoleDomain, Role Test
