--- src/main/java/com/t2m/g2nee/shop/bookset/Publisher/domain/Publisher.java
@@ -1,31 +1,0 @@
-package com.t2m.g2nee.shop.bookset.Publisher.domain;
-
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Entity
-@Table(name = "Publishers")
-@Getter
-@Setter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Publisher {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long publisherId;
-
-    private String publisherName;
-
-    private String publisherEngName;
-
-}
