--- src/main/java/com/t2m/g2nee/shop/bookset/Tag/repository/TagRepository.java
@@ -1,12 +1,0 @@
-package com.t2m.g2nee.shop.bookset.Tag.repository;
-
-import com.t2m.g2nee.shop.bookset.Publisher.domain.Publisher;
-import com.t2m.g2nee.shop.bookset.Tag.domain.Tag;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TagRepository extends JpaRepository<Tag,Long> {
-        Page<Tag> findAll(Pageable pageable);
-
-}
