+++ src/main/java/com/t2m/g2nee/shop/bookset/publisher/domain/Publisher.java
@@ -1,0 +1,31 @@
+package com.t2m.g2nee.shop.bookset.publisher.domain;
+
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.Table;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Entity
+@Table(name = "Publishers")
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Publisher {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long publisherId;
+
+    private String publisherName;
+
+    private String publisherEngName;
+
+}
