+++ src/main/java/com/t2m/g2nee/shop/bookset/publisher/mapper/PublisherMapper.java
@@ -1,0 +1,20 @@
+package com.t2m.g2nee.shop.bookset.publisher.mapper;
+
+import com.t2m.g2nee.shop.bookset.publisher.domain.Publisher;
+import com.t2m.g2nee.shop.bookset.publisher.dto.PublisherDto;
+import java.util.List;
+import org.mapstruct.Mapper;
+
+/**
+ * domain과 response를 변환하는 mapper 클래스
+ *
+ * @author : 신동민
+ * @since : 1.0
+ */
+@Mapper(componentModel = "spring")
+public interface PublisherMapper {
+
+    PublisherDto.Response entityToDto(Publisher publisher);
+
+    List<PublisherDto.Response> entitiesToDtos(List<Publisher> publisherList);
+}
