+++ src/main/java/com/t2m/g2nee/shop/bookset/tag/dto/TagDto.java
@@ -1,0 +1,39 @@
+package com.t2m.g2nee.shop.bookset.tag.dto;
+
+import com.fasterxml.jackson.annotation.JsonInclude;
+import javax.validation.constraints.NotBlank;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+public class TagDto {
+
+    private TagDto() {
+    }
+
+    @Getter
+    @Setter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    @Builder
+    public static class Request {
+
+        @NotBlank(message = "태그 이름을 입력해주세요.")
+        private String tagName;
+
+    }
+
+    @Getter
+    @Setter
+    @NoArgsConstructor
+    @AllArgsConstructor
+    @Builder
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    public static class Response {
+
+        private Long tagId;
+        private String tagName;
+    }
+}
