+++ src/main/java/com/t2m/g2nee/shop/bookset/tag/repository/TagRepository.java
@@ -1,0 +1,11 @@
+package com.t2m.g2nee.shop.bookset.tag.repository;
+
+import com.t2m.g2nee.shop.bookset.tag.domain.Tag;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TagRepository extends JpaRepository<Tag,Long> {
+        Page<Tag> findAll(Pageable pageable);
+
+}
