+++ src/main/resources/application.properties
@@ -19,7 +19,8 @@ spring.jpa.open-in-view=false
 # SecureKey properties
 nhncloud.key.url=https://api-keymanager.nhncloudservice.com/
 nhncloud.key.path=keymanager/v1.0/appkey/
-nhncloud.key.appKey=${APP_KEY}
+#nhncloud.key.appKey=${APP_KEY}
+nhncloud.key.appKey=500rMJPT021TnyNl
 nhncloud.key.usernameKeyId=/secrets/5465e99bf7204952b924dc2dcbec3355
 nhncloud.key.passwordKeyId=/secrets/f9a01c3e380f43afb0597b0447b753fc
 nhncloud.key.urlKeyId=/secrets/d8258f81c46946d2aa6ff294a61cfa3d
